# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Finatrades B2B Commodity Trade Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /b2b/auth/register:
    post:
      operationId: b2bRegister
      tags: [auth]
      summary: Register new B2B user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/B2BRegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /b2b/auth/login:
    post:
      operationId: b2bLogin
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /b2b/rfq:
    get:
      operationId: listRfqs
      tags: [rfq]
      summary: List my RFQs
      responses:
        "200":
          description: RFQ list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rfq"
    post:
      operationId: createRfq
      tags: [rfq]
      summary: Create Request for Quotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RfqInput"
      responses:
        "201":
          description: RFQ created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
  /b2b/rfq/{id}:
    get:
      operationId: getRfq
      tags: [rfq]
      summary: Get RFQ detail with offers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: RFQ detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"
  /b2b/rfq/{id}/offers:
    post:
      operationId: submitOffer
      tags: [rfq]
      summary: Seller submits offer on RFQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Offer submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /b2b/rfq/{id}/accept:
    post:
      operationId: acceptOffer
      tags: [rfq]
      summary: Buyer accepts an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptOfferInput"
      responses:
        "200":
          description: Offer accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rfq"

  # ─── ORDERS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    B2BRegisterInput:
      type: object
      required: [email, password, fullName, role, phone]
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [exporter, importer, government]
        phone:
          type: string
        companyName:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/B2BUser"
        token:
          type: string
        message:
          type: string
    B2BUser:
      type: object
      required: [id, email, fullName, role, onboardingStep, onboardingComplete]
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [exporter, importer, government]
        companyName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        onboardingStep:
          type: integer
        onboardingComplete:
          type: boolean
        kycStatus:
          type: string
          enum: [pending, under_review, approved, rejected]
        createdAt:
          type: string

    # ── ONBOARDING ──
...
    RfqInput:
      type: object
      required: [commodityType, quantity, unit, hubId, targetPricePerUnit, currency, deliveryTerms, expiresAt]
      properties:
        commodityType:
          type: string
        grade:
          type: string
        quantity:
          type: number
        unit:
          type: string
        hubId:
          type: string
        targetPricePerUnit:
          type: number
        currency:
          type: string
        deliveryTerms:
          type: string
        attachments:
          type: string
        expiresAt:
          type: string
        notes:
          type: string
    OfferInput:
      type: object
      required: [pricePerUnit, currency, availableQuantity, leadTimeDays]
      properties:
        pricePerUnit:
          type: number
        currency:
          type: string
        availableQuantity:
          type: number
        leadTimeDays:
          type: integer
        inventoryId:
          type: string
        terms:
          type: string
        validUntil:
          type: string
    AcceptOfferInput:
      type: object
      required: [offerId]
      properties:
        offerId:
          type: string
    Offer:
      type: object
      required: [id, rfqId, sellerId, sellerName, pricePerUnit, currency, availableQuantity, status, createdAt]
      properties:
        id:
          type: string
        rfqId:
          type: string
        sellerId:
          type: string
        sellerName:
          type: string
        pricePerUnit:
          type: number
        currency:
          type: string
        availableQuantity:
          type: number
        leadTimeDays:
          type: integer
        inventoryId:
          type: ["string", "null"]
        terms:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected, expired]
        createdAt:
          type: string
    Rfq:
      type: object
      required: [id, rfqRef, buyerId, buyerName, commodityType, quantity, unit, hubId, status, createdAt]
      properties:
        id:
          type: string
        rfqRef:
          type: string
        buyerId:
          type: string
        buyerName:
          type: string
        commodityType:
          type: string
        grade:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        hubId:
          type: string
        hubName:
          type: string
        targetPricePerUnit:
          type: ["number", "null"]
        currency:
          type: string
        deliveryTerms:
          type: string
        status:
          type: string
          enum: [open, offers_received, negotiating, confirmed, converted, expired, cancelled]
        offers:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── ORDERS ──
...